Casts time values to uint32_t for consistency
Ensures consistent type usage for time-related variables by casting them to `uint32_t`.

This avoids potential implicit conversions and ensures data integrity when dealing with time values,
especially when interacting with timer functions.

+++ PAC/common/cip_tech_def.h
@@ -481,8 +481,8 @@ class MSAPID
         float q2;
         float Uk;
         float dUk;
-        unsigned long start_time;
-        unsigned long last_time;
+        uint32_t start_time;
+        uint32_t last_time;
 
         char prev_manual_mode;
